Add endpoint for getting seats active to reserve

+++ src/main/java/com/example/mock/entity/Seat.java
@@ -18,7 +18,7 @@ public class Seat {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public long seatId;
+    public Integer seatId;
 
     public String seatRow;
     public int seatNumber;
